Add an offline first architecture for the main list of movies.
Create a local storage for movies using a local database with Room and DataStore.
The movies coming from the database are now the source of truth for the main list of movies.
Also add the appropriate repositories and datasources to manage the different sources of data: remote API and local storage.
In addition, the code of the application has been reorganised and some names of variables, methods and classes have been renamed to adapt them to the offline first architecture.

+++ app/src/main/java/com/example/mymovies/domain/usecases/GetMovieDetailsUseCase.kt
@@ -2,10 +2,10 @@ package com.example.mymovies.domain.usecases
 
 import com.example.mymovies.data.repositories.MovieDetailsRepository
 import com.example.mymovies.domain.models.MovieDetails
-import com.example.mymovies.domain.models.toDomainModel
+import com.example.mymovies.data.remote.models.toDomainModel
 
 class GetMovieDetailsUseCase(private val repository: MovieDetailsRepository) {
 
-    suspend operator fun invoke(movieId: Int): MovieDetails =
+    suspend operator fun invoke(movieId: Long): MovieDetails =
         repository.getMovieDetails(movieId).toDomainModel()
 }
